requirements.txt
Django==4.2.0
plyer==2.1
requests==2.31.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
python-dotenv==1.0.1
whitenoise==6.8.2
gunicorn==23.0.0